fix coco processor

+++ data/processing/coco_processor.py
@@ -4,7 +4,7 @@ from pathlib import Path
 import pycocotools.coco as coco
 import pandas as pd
 from pyspark.sql import SparkSession
-from pyspark.sql.types import StructType, StructField, StringType, ArrayType, FloatType, IntegerType
+from pyspark.sql.types import StructType, StructField, StringType, ArrayType, FloatType, IntegerType, LongType
 from ..unity_catalog.catalog_manager import CatalogManager
 
 class COCOProcessor:
@@ -56,13 +56,13 @@ class COCOProcessor:
     def create_spark_dataframe(self, df: pd.DataFrame) -> 'pyspark.sql.DataFrame':
         """Convert pandas DataFrame to Spark DataFrame with proper schema."""
         schema = StructType([
-            StructField('image_id', IntegerType(), False),
+            StructField('image_id', LongType(), False),
             StructField('file_name', StringType(), False),
             StructField('width', IntegerType(), False),
             StructField('height', IntegerType(), False),
             StructField('annotations', ArrayType(
                 StructType([
-                    StructField('id', IntegerType(), False),
+                    StructField('id', LongType(), False),
                     StructField('category_id', IntegerType(), False),
                     StructField('bbox', ArrayType(FloatType()), False),
                     StructField('segmentation', ArrayType(ArrayType(FloatType())), False),
